Moved Transaction component and it's AddTransaction logic into seperate components. Also added some type of delegate and anonymous function usage
